+++ Models/TransaccionActualizacionViewModel.cs
@@ -3,7 +3,7 @@
     public class TransaccionActualizacionViewModel : TransaccionCreacionViewModel
     {
         public int CuentaAnteriorId { get; set; }
-        public double MontoAnterior { get; set; }
+        public int MontoAnterior { get; set; }
         public string UrlRetorno { get; set; }
     }
 }
